move functions to separate utility file
