+++ addon/mixins/sortable-item.js
@@ -1,8 +1,10 @@
 import Ember from 'ember';
 import computed from 'ember-new-computed';
+import {invokeAction} from 'ember-invoke-action';
+
 import scrollParent from '../system/scroll-parent';
 import ScrollContainer from '../system/scroll-container';
-import {invokeAction} from 'ember-invoke-action';
+import {getX, getY, getBorderSpacing} from '../system/utils';
 
 const { Mixin, $, run } = Ember;
 const { Promise } = Ember.RSVP;
@@ -671,60 +673,3 @@ export default Mixin.create({
     this._tellGroup('commit');
   }
 });
-
-/**
-  Gets the y offset for a given event.
-  Work for touch and mouse events.
-  @method getY
-  @return {Number}
-  @private
-*/
-function getY(event) {
-  let originalEvent = event.originalEvent;
-  let touches = originalEvent && originalEvent.changedTouches;
-  let touch = touches && touches[0];
-
-  if (touch) {
-    return touch.screenY;
-  } else {
-    return event.pageY;
-  }
-}
-
-/**
-  Gets the x offset for a given event.
-  @method getX
-  @return {Number}
-  @private
-*/
-function getX(event) {
-  let originalEvent = event.originalEvent;
-  let touches = originalEvent && originalEvent.changedTouches;
-  let touch = touches && touches[0];
-
-  if (touch) {
-    return touch.screenX;
-  } else {
-    return event.pageX;
-  }
-}
-
-/**
-  Gets a numeric border-spacing values for a given element.
-
-  @method getBorderSpacing
-  @param {Element} element
-  @return {Object}
-  @private
-*/
-function getBorderSpacing(el) {
-  el = $(el);
-
-  let css = el.css('border-spacing'); // '0px 0px'
-  let [horizontal, vertical] = css.split(' ');
-
-  return {
-    horizontal: parseFloat(horizontal),
-    vertical: parseFloat(vertical)
-  };
-}
